Task2. Build Variants and code separation
