+++ TestsLesson4/app/src/main/java/com/geekbrains/tests/presenter/RepositoryContract.kt
@@ -2,7 +2,7 @@ package com.geekbrains.tests.presenter
 
 import com.geekbrains.tests.repository.RepositoryCallback
 
-internal interface RepositoryContract {
+interface RepositoryContract {
     fun searchGithub(
         query: String,
         callback: RepositoryCallback
